+++ libdbg/csharp/PS4DBG.Kernel.cs
@@ -1,0 +1,58 @@
+using System;
+
+namespace libdebug
+{
+    public partial class PS4DBG
+    {
+        // kernel
+        //packet sizes
+        //send size
+        private const int CMD_KERN_READ_PACKET_SIZE = 12;
+        private const int CMD_KERN_WRITE_PACKET_SIZE = 12;
+        //receive size
+        private const int KERN_BASE_SIZE = 8;
+
+
+        /// <summary>
+        /// Get kernel base address
+        /// </summary>
+        /// <returns></returns>
+        public ulong KernelBase()
+        {
+            CheckConnected();
+
+            SendCMDPacket(CMDS.CMD_KERN_BASE, 0);
+            CheckStatus();
+            return BitConverter.ToUInt64(ReceiveData(KERN_BASE_SIZE), 0);
+        }
+
+        /// <summary>
+        /// Read memory from kernel
+        /// </summary>
+        /// <param name="address">Memory address</param>
+        /// <param name="length">Data length</param>
+        /// <returns></returns>
+        public byte[] KernelReadMemory(ulong address, int length)
+        {
+            CheckConnected();
+
+            SendCMDPacket(CMDS.CMD_KERN_READ, CMD_KERN_READ_PACKET_SIZE, address, length);
+            CheckStatus();
+            return ReceiveData(length);
+        }
+
+        /// <summary>
+        /// Write memory in kernel
+        /// </summary>
+        /// <param name="address">Memory address</param>
+        /// <param name="data">Data</param>
+        public void KernelWriteMemory(ulong address, byte[] data)
+        {
+            CheckConnected();
+
+            SendCMDPacket(CMDS.CMD_KERN_WRITE, CMD_KERN_WRITE_PACKET_SIZE, address, data.Length);
+            SendData(data, data.Length);
+            CheckStatus();
+        }
+    }
+}
